Make Tailcat demo flows easier to follow
